+++ thriftrw/compile/scope.py
@@ -103,6 +103,7 @@ class Scope(object):
         self.const_specs[const_spec.name] = const_spec
 
     def add_surface(self, name, surface):
+        """Adds a top-level attribute with the given name to the module."""
         assert surface is not None
 
         if hasattr(self.module, name):
@@ -111,23 +112,6 @@ class Scope(object):
             )
 
         setattr(self.module, name, surface)
-
-    def add_function(self, name, func):
-        """Adds a top-level function with the given name to the module.
-
-        :param name:
-            Name of the function.
-        :param func:
-            Function to add to the module.
-        """
-        assert func is not None
-
-        if hasattr(self.module, name):
-            raise ThriftCompilerError(
-                'Cannot define "%s". The name has already been used.' % name
-            )
-
-        setattr(self.module, name, func)
 
     def add_type_spec(self, name, spec, lineno):
         """Adds the given type to the scope.