+++ thriftrw/protocol/binary.pyx
@@ -21,7 +21,6 @@
 from __future__ import absolute_import, unicode_literals, print_function
 
 import struct
-from io import BytesIO
 from six.moves import range
 
 from libc.stdint cimport (
@@ -32,6 +31,9 @@ from libc.stdint cimport (
 )
 
 from thriftrw.wire cimport ttype
+from thriftrw.wire.message cimport Message
+from thriftrw.wire.mtype import MESSAGE_TYPES
+
 from thriftrw._buffer cimport ReadBuffer
 from thriftrw._buffer cimport WriteBuffer
 from thriftrw.wire.value cimport (
@@ -52,7 +54,7 @@ from thriftrw.wire.value cimport (
     ListValue,
 )
 
-from .core import Protocol
+from .core cimport Protocol
 from ..errors import EndOfInputError
 from ..errors import ThriftProtocolError
 from ._endian cimport (
@@ -67,6 +69,18 @@ from ._endian cimport (
 
 STRUCT_END = 0
 
+VERSION = 1
+
+# Under strict mode, the version number is in most significant 16-bits of the
+# 32-bit integer, and the most significant bit is set. This mask gets &-ed
+# with the integer to get just the version number (it still needs to be
+# shifted 16-bits to the right).
+VERSION_MASK = 0x7fff0000
+
+# The least significant 8-bits of the 32-bit integer contain the type.
+TYPE_MASK = 0x000000ff
+
+
 cdef class BinaryProtocolReader(object):
     """Parser for the binary protocol."""
 
@@ -77,7 +91,8 @@ cdef class BinaryProtocolReader(object):
 
         :param reader:
             File-like object with a ``read(num)`` method which returns
-            *exactly* the requested number of bytes.
+            *exactly* the requested number of bytes, or all remaining bytes if
+            ``num`` is negative.
         """
         self.reader = reader
 
@@ -136,6 +151,37 @@ cdef class BinaryProtocolReader(object):
     cdef double _double(self) except *:
         cdef int64_t value = self._i64()
         return (<double*>(&value))[0]
+
+    cdef Message read_message(self):
+        size = self._i32()
+        # TODO with cython, some of the Python-specific hacks around bit
+        # twiddling can be skipped.
+        if size < 0:
+            # strict version:
+            #
+            #     versionAndType:4 name~4 seqid:4 payload
+            version = (size & VERSION_MASK) >> 16
+            if version != VERSION:
+                raise ThriftProtocolError(
+                    'Unsupported version "%r"' % version
+                )
+            typ = size & TYPE_MASK
+            size = self._i32()
+            name = self.reader.take(size)
+        else:
+            # non-strict version:
+            #
+            #     name:4 type:1 seqid:4 payload
+            name = self.reader.take(size)
+            typ = self._byte()
+
+        if typ not in MESSAGE_TYPES:
+            raise ThriftProtocolError('Unknown message type "%r"' % typ)
+
+        seqid = self._i32()
+        body = self.read(ttype.STRUCT)
+
+        return Message(name=name, seqid=seqid, body=body, message_type=typ)
 
     cdef BoolValue read_bool(self):
         """Reads a boolean."""
@@ -256,7 +302,7 @@ cdef class BinaryProtocolWriter(ValueVisitor):
         """
         self.writer = writer
 
-    cpdef write(self, value):
+    cpdef void write(self, Value value):
         """Writes the given value.
 
         :param value:
@@ -334,23 +380,32 @@ cdef class BinaryProtocolWriter(ValueVisitor):
         for v in values:
             self.write(v)
 
+    cdef void write_message(self, Message message) except *:
+        self.visit_binary(message.name)
+        self.visit_byte(message.message_type)
+        self.visit_i32(message.seqid)
+        self.write(message.body)
 
-class BinaryProtocol(Protocol):
+
+cdef class BinaryProtocol(Protocol):
     """Implements the Thrift binary protocol."""
 
-    __slots__ = ()
-
-    writer_class = BinaryProtocolWriter
-    reader_class = BinaryProtocolReader
-
-    def serialize_value(self, value):
-        buff = WriteBuffer()
-        self.writer_class(buff).write(value)
+    cpdef bytes serialize_message(self, Message message):
+        cdef WriteBuffer buff = WriteBuffer()
+        BinaryProtocolWriter(buff).write_message(message)
         return buff.value
 
-    def deserialize_value(self, typ, s):
-        buff = ReadBuffer(s)
-        reader = self.reader_class(buff)
-        return reader.read(typ)
+    cpdef Message deserialize_message(self, bytes s):
+        cdef ReadBuffer buff = ReadBuffer(s)
+        return BinaryProtocolReader(buff).read_message()
+
+    cpdef bytes serialize_value(self, Value value):
+        cdef WriteBuffer buff = WriteBuffer()
+        BinaryProtocolWriter(buff).write(value)
+        return buff.value
+
+    cpdef Value deserialize_value(self, int typ, bytes s):
+        cdef ReadBuffer buff = ReadBuffer(s)
+        return BinaryProtocolReader(buff).read(typ)
 
 __all__ = ['BinaryProtocol']